Auto interpolate menu item routes with given data, move menu item utils into their own file
