+++ client-v2/src/app/pages/home/entity-page/entity-page.component.ts
@@ -4,9 +4,10 @@ import { UntilDestroy } from '@ngneat/until-destroy'
 import { Store } from '@ngrx/store'
 import { map, of, shareReplay, switchMap } from 'rxjs'
 import { Breadcrumb } from 'src/app/components/molecules/breadcrumbs/breadcrumbs.component'
-import { getEntityMenuItemsMap, useDataForAction } from 'src/app/shared/entity-menu-items'
+import { getEntityMenuItemsMap } from 'src/app/shared/entity-menu-items'
 import { AppState } from 'src/app/store'
 import { traceEntity, traceEntityIncludingTasks } from 'src/app/store/entities/utils'
+import { useDataForAction } from 'src/app/utils/menu-item.helpers'
 
 @UntilDestroy()
 @Component({
